+++ packages/react-components/tests/browser/RichTextSendBox.spec.tsx
@@ -22,12 +22,12 @@ betaTest.describe('RichTextSendBox tests', () => {
     await expect(component).toHaveScreenshot('rich-text-send-box-without-format-toolbar.png');
     const formatButton = component.getByTestId(formatButtonId);
 
-    await formatButton.hover();
+    await component.getByTestId('rooster-rich-text-editor').hover();
     await expect(component).toHaveScreenshot('rich-text-send-box-hover.png');
 
     await formatButton.click();
-    //move mouse to the format button so the screenshots are consistent
-    await formatButton.hover();
+    //move mouse to the editor so the screenshots are consistent
+    await component.getByTestId('rooster-rich-text-editor').hover();
     await expect(component).toHaveScreenshot('rich-text-send-box-with-format-toolbar.png');
   });
